create BufferingDetector
